% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_FI.R
\name{AUC_FI}
\alias{AUC_FI}
\title{Compute AUC (and ROC) when verification sample selected by preliminary score, Full Imputation Method}
\usage{
AUC_FI(dat, model = disease ~ qlogis(score), scrvar = "score")
}
\arguments{
\item{dat}{data.frame containing model variables, including scrvar}

\item{model}{Model to estimate selection to verification sample based on screening score}

\item{scrvar}{Name of screening variable within dat}
}
\value{
Area under the ROC curve estimated using Full Imputation method (see Alonzo and Pepe, 2003). Also up to 100 points of ROC curve.
}
\description{
Compute AUC (and ROC) when verification sample selected by preliminary score, Full Imputation Method
}
\examples{
# none yet
}
