% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid.R
\name{trapezoid}
\alias{trapezoid}
\title{Compute area by trapezoid rule}
\usage{
trapezoid(x, y)
}
\arguments{
\item{x}{horizontal coordinates}

\item{y}{vertical coordinates}
}
\value{
area under curve
}
\description{
Compute area by trapezoid rule
}
\examples{
trapezoid(seq(0, 10), rnorm(11, 5, 1))
}
