% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCverification.R
\name{ROCverification}
\alias{ROCverification}
\title{Estimation of ROC in the presence of verification bias}
\usage{
ROCverification(
  data = NULL,
  dvar = NULL,
  tvar = NULL,
  estimators = c("cc", "fi", "msi", "ipw", "dr"),
  dformula = NULL,
  vformula = NULL,
  vvar = NULL,
  cc = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{dvar}{name of disease variable}

\item{tvar}{name of scoring variable}

\item{estimators}{which (of 5) estimators to return}

\item{dformula}{formula for d given t and possibly x}

\item{vformula}{formula for v given t and possibly x}

\item{vvar}{name of selection variable}

\item{cc}{thresholds at which to estimate TPR and FPR}
}
\value{
data.frame with cc, TPR and FPR for requested estimators
}
\description{
Estimation of ROC in the presence of verification bias
}
\examples{
# None yet
}
